requirements.txt
alembic==1.13.1
Flask~=3.0.0
flask_cors==5.0.1
Flask_JWT_Extended==4.7.1
Flask_Migrate==3.1.0
flask_sqlalchemy==3.1.1
grpcio==1.71.0
numpy~=1.26.4
opencv_contrib_python==4.11.0.86
opencv_python==4.10.0.84
protobuf==6.31.1
qrcode==7.4.2
SQLAlchemy==2.0.23
transliterate==1.10.2
psycopg2-binary==2.9.10
Pillow==10.1.0